+++ src/rules/function_naming.zig
@@ -201,8 +201,7 @@ fn run(
 
     return if (lint_problems.items.len > 0)
         try zlinter.results.LintResult.init(
-            session_arena,
-            doc.absPath(session),
+            doc.file_id,
             lint_problems.items,
         )
     else
@@ -309,7 +308,7 @@ fn allocResolvedDeclNotes(
         1,
     );
     notes[0] = .{
-        .abs_path = try session_arena.dupe(u8, decl_location.abs_path),
+        .file_id = decl_location.file_id,
         .start = decl_location.start,
         .end = decl_location.end,
         .line = decl_location.line,